Support date range for SSID Request extract
This pattern is a little different than the others. Sets the variables
for xpaths or IDs in a conditional rather than putting the entire code
block in a `if not by_date_range` conditional. Looks cleaner and is
probably the best way to do it. Might apply it to others later.