Functional ODS Downloads are upon us
Was able to get an ODS download to work. Additionally, it would appear
the ODS and Snapshot have a lot more in common in how they are
processed. I can generalize for both the Snapshot and the ODS and open
up MORE if not ALL field options for downloading from the Snapshots as
well. Unbelievable.